+++ src/JustEat.StatsD/Transport/DataExtensions.cs
@@ -1,0 +1,25 @@
+using System;
+
+namespace JustEat.StatsD
+{
+    public static class DataExtensions
+    {
+        public static ReadOnlySpan<byte> GetSpan(this in Data src)
+        {
+            if (src.Array == null)
+            {
+                return src.Span;
+            }
+            return src.Array.AsSpan();
+        }
+
+        public static byte[] GetArray(this in Data src)
+        {
+            if (src.Array == null)
+            {
+                return src.Span.ToArray();
+            }
+            return src.Array;
+        }
+    }
+}
